backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
curl_cffi==0.13.0
dj-database-url==3.0.1
Django==4.2.30
django-cors-headers==4.9.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
frozendict==2.4.7
idna==3.11
kombu==5.6.2
multitasking==0.0.12
numpy==2.0.2
packaging==26.0
pandas==2.3.3
peewee==4.0.4
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.33.6
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.12.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
redis==7.0.1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
yfinance==1.2.0
drf-spectacular==0.28.0
openpyxl==3.1.5
